requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.7.0
gunicorn==23.0.0
whitenoise==6.9.0
